+++ CHANGELOG.md
@@ -175,6 +175,7 @@ Modifications can only be done via builders.
 - Work on shouldStopProcessingAfter
 - Rename shouldDigInto into shouldStopProcessingAfter
 - Provide a method to stop processing at a certain point
+- Implementers of ScBuilderBluePrint must implement shouldStopProcessingAfter
 
 ### Fixed